+++ packages/consent/consent-tools/src/domain/consent-stamping.ts
@@ -1,7 +1,7 @@
 import { AnyAnalytics, Categories } from '../types'
 
 type CreateConsentMw = (
-  getCategories: () => Promise<Categories | undefined>
+  getCategories: () => Promise<Categories>
 ) => AnyAnalytics['addSourceMiddleware']
 
 /**
@@ -11,10 +11,6 @@ export const createConsentStampingMiddleware: CreateConsentMw =
   (getCategories) =>
   async ({ payload, next }) => {
     const categories = await getCategories()
-    if (!categories) {
-      console.error('Skipping consent stamping because categories are empty')
-      return next(payload)
-    }
     payload.obj.context.consent = {
       ...payload.obj.context.consent,
       categoryPreferences: categories,