+++ packages/consent/consent-tools/src/types/settings.ts
@@ -50,41 +50,40 @@ export interface CreateWrapperSettings {
   integrationCategoryMappings?: IntegrationCategoryMappings
 
   /**
-   * Predicate function to override default logic around whether or not to load an integration.
-   * @default
-   * ```ts
-   * // consent if user consents to at least one category defined in the integration
-   * (integrationCategories, categories, _info) => {
-   *    if (!integrationCategories.length) return true
-   *    return integrationCategories.some((c) => categories[c])
-   * }
-   * ```
-   *
-   * @example -
-   * ```ts
-   * (integrationCategories, categories, _info) => {
-   * // consent if user consents to _all_ categories defined in the integration
-   *    if (!integrationCategories.length) return true
-   *    return integrationCategories.every((c) => categories[c])
-   * }
-   * ```
-   *
+   * Predicate function to override default logic around whether or not to load an integration. By default, consent requires a user to have all categories enabled for a given integration.
    * @example
    * ```ts
-   * // count consent as usual, but always disable a particular plugin
-   * (integrationCategories, categories, { creationName }) => {
+   * // Always disable a particular plugin
+   * const shouldEnableIntegration = (integrationCategories, categories, { creationName }) => {
    *    if (creationName === 'FullStory') return false
    *    if (!integrationCategories.length) return true
-   *    return integrationCategories.some((c) => categories[c])
+   *    return integrationCategories.every((c) => categories[c])
    * }
    * ```
    */
   shouldEnableIntegration?: (
     integrationCategories: string[],
     categories: Categories,
-    integrationInfo: Pick<
-      CDNSettingsRemotePlugin,
-      'creationName' | 'libraryName'
-    >
+    integrationInfo: Pick<CDNSettingsRemotePlugin, 'creationName'>
   ) => boolean
+
+  /**
+   * Prune consent categories from the `context.consent.categoryPreferences` payload if that category is not mapped to any integration in your Segment.io source.
+   * This is helpful if you want to save on bytes sent to Segment and do need the complete list of CMP's categories for debugging or other reasons.
+   * By default, all consent categories returned by `getCategories()` are sent to Segment.
+   * @default false
+   * ### Example Behavior
+   * You have the following categories mappings defined:
+   * ```
+   * FullStory -> 'CAT002',
+   * Braze -> 'CAT003'
+   * ```
+   * ```ts
+   * // pruneUnmappedCategories = false (default)
+   * { CAT0001: true, CAT0002: true, CAT0003: true }
+   * // pruneUnmappedCategories = true
+   * { CAT0002: true, CAT0003: true  } // pruneUnmappedCategories = true
+   * ```
+   */
+  pruneUnmappedCategories?: boolean
 }