+++ packages/browser/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @segment/analytics-next
 
+## 1.53.3
+
+### Patch Changes
+
+- [#915](https://github.com/segmentio/analytics-next/pull/915) [`7072377`](https://github.com/segmentio/analytics-next/commit/7072377c0adae54b844a441fa545763e0ff654ba) Thanks [@danieljackins](https://github.com/danieljackins)! - Fix batching after page navigation
+
+## 1.53.2
+
+### Patch Changes
+
+- [#888](https://github.com/segmentio/analytics-next/pull/888) [`f3183f2`](https://github.com/segmentio/analytics-next/commit/f3183f2163d93f98ce2f8dd7830d2bca3b47f537) Thanks [@danieljackins](https://github.com/danieljackins)! - Fix query string parsing bug that was causing events containing the 'search' property with a non string value to be dropped
+
+* [#891](https://github.com/segmentio/analytics-next/pull/891) [`e0c7792`](https://github.com/segmentio/analytics-next/commit/e0c7792924e6f7ba0a36c200c5468141de3b9320) Thanks [@zikaari](https://github.com/zikaari)! - Fix cookie write error
+
+## 1.53.1
+
+### Patch Changes
+
+- [#900](https://github.com/segmentio/analytics-next/pull/900) [`9c8b609`](https://github.com/segmentio/analytics-next/commit/9c8b609f500d2338fdbf2572639241ffdfd27f8f) Thanks [@silesky](https://github.com/silesky)! - Add consent info to typescript types
+
+* [#896](https://github.com/segmentio/analytics-next/pull/896) [`48ce3ec`](https://github.com/segmentio/analytics-next/commit/48ce3ecbdea9361dde90481d7c50207613602d52) Thanks [@oscb](https://github.com/oscb)! - Added support for `screen` events for Segment destination
+
 ## 1.53.0
 
 ### Minor Changes